[FIX] #277 비관적 락 적용
